Extract utility functions from `CurrentSurvey` component.
