+++ assets/js/components/surveys/CurrentSurvey/constants.js
@@ -1,0 +1,22 @@
+/**
+ * CurrentSurvey component constants.
+ *
+ * Site Kit by Google, Copyright 2025 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export const TYPE_MULTI_SELECT = 'multi_select';
+export const TYPE_OPEN_TEXT = 'open_text';
+export const TYPE_RATING = 'rating';
+export const TYPE_SINGLE_SELECT = 'single_select';
